[leetcode/ac] 278. First Bad Version
Runtime: 0 ms, faster than 100.00% of C++ online submissions for First Bad Version.
Memory Usage: 5.7 MB, less than 99.72% of C++ online submissions for First Bad Version.

+++ leetcode/solved/278.cc
@@ -4,8 +4,8 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        int low = 1;
-        int high = n;
+        long low = 1;
+        long high = n;
         
         while(low<high) {
             const int version = (low+high) / 2;
